+++ tests/test_deserialize_sequence.cc
@@ -4,7 +4,7 @@
  * \date   Created on July 26, 2023
  * \brief  Test suite for free functions declared in deserialize_sequence.h.
  *
- * \copyright Copyright 2023 Deutsches Elektronen-Synchrotron (DESY), Hamburg
+ * \copyright Copyright 2023-2024 Deutsches Elektronen-Synchrotron (DESY), Hamburg
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published
@@ -28,13 +28,3 @@
 
 using namespace task;
 using namespace task::literals;
-
-TEST_CASE("get_sequence_info_from_filename()", "[deserialize_sequence]")
-{
-    REQUIRE(get_sequence_info_from_filename("my_sequence[deadbeef]")
-        == SequenceInfo{ "", SequenceName{ "my_sequence" }, 0xdeadbeef_uid });
-    REQUIRE(get_sequence_info_from_filename("[1234567890abcdef]")
-        == SequenceInfo{ "", gul14::nullopt, 0x1234567890abcdef_uid });
-    REQUIRE(get_sequence_info_from_filename("A$2f$22sequence$22$24$3cagain$3e [my_uid]")
-        == SequenceInfo{ "A/\"sequence\"$<again> [my_uid]", gul14::nullopt, gul14::nullopt });
-}
